requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorzero==2.0
gast==0.2.2
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.1.8
gpiozero==1.5.1
grpcio==1.27.2
h5py==2.10.0
idna==2.10
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
mutagen==1.45.1
numpy==1.18.1
oauthlib==3.1.0
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
opt-einsum==3.3.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.0.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.14.0
tensorboard==2.0.2
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.26.4
Werkzeug==1.0.0
wrapt==1.11.2
